*RuneCostID corrected *CMSG_GROUP_INVITE updated *Pet bar corrected *Pet spells now works... (but server still crashes when use spells) *Professions max skill now is 450

+++ src/arcemu-world/Player.h
@@ -294,7 +294,7 @@ struct CreateInfo_SkillStruct
 struct CreateInfo_ActionBarStruct
 {
 	uint32  button;
-	uint32  action;
+	uint16  action;
 	uint32  type;
 	uint32  misc;
 };
